sprint_8/project/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nikolay Stepanov <nikolai.d.stepanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
